Update usage of Shout
Shout implements code to deal with fonts but this code is never used and probably do not work with Pharo's scaling and other features. 
To simplify Shout this will probably be removed leading to the removal of #initialTextAttributesForPixelHeight:. 

Since Microdown is using this method, I propose an alternative way to update the behavior of Microdown styler to allow the refactor of Shout.