fix: allow MicCSSEntity to overwrite existing file
When telling its CSS file to copy itself to a destination directory, MicCSSEntity was giving its MicCSSFile the full path, including the file basename. If the file alredy existed at that path, the MicCSSFile was attempting to delete the full path and appending the basename again, leading to an infinite loop.

This completely moves the responsibility for knowing the basename to MicCSSFile.